+++ plugin/evm/header/base_fee.go
@@ -5,6 +5,7 @@ package header
 
 import (
 	"errors"
+	"fmt"
 	"math/big"
 
 	"github.com/ava-labs/coreth/core/types"
@@ -17,8 +18,19 @@ var errEstimateBaseFeeWithoutActivation = errors.New("cannot estimate base fee f
 // calculates the expected base fee for the child block.
 //
 // Prior to AP3, the returned base fee will be nil.
-func BaseFee(config *params.ChainConfig, parent *types.Header, timestamp uint64) (*big.Int, error) {
+func BaseFee(
+	config *params.ChainConfig,
+	parent *types.Header,
+	timestamp uint64,
+) (*big.Int, error) {
 	switch {
+	case config.IsFUpgrade(timestamp):
+		state, err := feeStateBeforeBlock(config, parent, timestamp)
+		if err != nil {
+			return nil, fmt.Errorf("calculating initial fee state: %w", err)
+		}
+		price := state.GasPrice()
+		return new(big.Int).SetUint64(uint64(price)), nil
 	case config.IsApricotPhase3(timestamp):
 		return baseFeeFromWindow(config, parent, timestamp)
 	default:
@@ -35,7 +47,11 @@ func BaseFee(config *params.ChainConfig, parent *types.Header, timestamp uint64)
 //
 // Warning: This function should only be used in estimation and should not be
 // used when calculating the canonical base fee for a block.
-func EstimateNextBaseFee(config *params.ChainConfig, parent *types.Header, timestamp uint64) (*big.Int, error) {
+func EstimateNextBaseFee(
+	config *params.ChainConfig,
+	parent *types.Header,
+	timestamp uint64,
+) (*big.Int, error) {
 	if config.ApricotPhase3BlockTimestamp == nil {
 		return nil, errEstimateBaseFeeWithoutActivation
 	}