+++ plugin/evm/header/dynamic_fee_windower.go
@@ -4,35 +4,27 @@
 package header
 
 import (
-	"encoding/binary"
 	"errors"
 	"fmt"
 	"math/big"
 
-	"github.com/ava-labs/avalanchego/utils/wrappers"
 	"github.com/ava-labs/coreth/core/types"
 	"github.com/ava-labs/coreth/params"
-	"github.com/ava-labs/coreth/plugin/evm/ap3"
-	"github.com/ava-labs/coreth/plugin/evm/ap4"
-	"github.com/ava-labs/coreth/plugin/evm/ap5"
+	"github.com/ava-labs/coreth/plugin/evm/upgrade/ap3"
+	"github.com/ava-labs/coreth/plugin/evm/upgrade/ap4"
+	"github.com/ava-labs/coreth/plugin/evm/upgrade/ap5"
+	"github.com/ava-labs/coreth/plugin/evm/upgrade/etna"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/math"
 )
-
-// FeeWindowSize is the number of bytes that are used to encode the dynamic fee
-// window in the header's Extra field after the Apricot Phase 3 upgrade.
-const FeeWindowSize = wrappers.LongLen * ap3.WindowLen
 
 var (
 	maxUint256Plus1 = new(big.Int).Lsh(common.Big1, 256)
 	maxUint256      = new(big.Int).Sub(maxUint256Plus1, common.Big1)
 
-	ap3MinBaseFee = big.NewInt(ap3.MinBaseFee)
-	ap4MinBaseFee = big.NewInt(ap4.MinBaseFee)
-	// EtnaMinBaseFee is exported so that it can be modified by tests.
-	//
-	// TODO: Unexport this.
-	EtnaMinBaseFee = big.NewInt(params.EtnaMinBaseFee)
+	ap3MinBaseFee  = big.NewInt(ap3.MinBaseFee)
+	ap4MinBaseFee  = big.NewInt(ap4.MinBaseFee)
+	etnaMinBaseFee = big.NewInt(etna.MinBaseFee)
 
 	ap3MaxBaseFee = big.NewInt(ap3.MaxBaseFee)
 	ap4MaxBaseFee = big.NewInt(ap4.MaxBaseFee)
@@ -40,7 +32,7 @@ var (
 	ap3BaseFeeChangeDenominator = big.NewInt(ap3.BaseFeeChangeDenominator)
 	ap5BaseFeeChangeDenominator = big.NewInt(ap5.BaseFeeChangeDenominator)
 
-	errDynamicFeeWindowInsufficientLength = errors.New("insufficient length for dynamic fee window")
+	errInvalidTimestamp = errors.New("invalid timestamp")
 )
 
 // baseFeeFromWindow should only be called if `timestamp` >= `config.ApricotPhase3Timestamp`
@@ -136,7 +128,7 @@ func baseFeeFromWindow(config *params.ChainConfig, parent *types.Header, timesta
 	// Ensure that the base fee does not increase/decrease outside of the bounds
 	switch {
 	case config.IsEtna(parent.Time):
-		baseFee = selectBigWithinBounds(EtnaMinBaseFee, baseFee, maxUint256)
+		baseFee = selectBigWithinBounds(etnaMinBaseFee, baseFee, maxUint256)
 	case isApricotPhase5:
 		baseFee = selectBigWithinBounds(ap4MinBaseFee, baseFee, maxUint256)
 	case config.IsApricotPhase4(parent.Time):
@@ -163,13 +155,14 @@ func feeWindow(
 		return ap3.Window{}, nil
 	}
 
-	dynamicFeeWindow, err := parseFeeWindow(parent.Extra)
+	dynamicFeeWindow, err := ap3.ParseWindow(parent.Extra)
 	if err != nil {
 		return ap3.Window{}, err
 	}
 
 	if timestamp < parent.Time {
-		return ap3.Window{}, fmt.Errorf("cannot calculate fee window for timestamp %d prior to parent timestamp %d",
+		return ap3.Window{}, fmt.Errorf("%w: timestamp %d prior to parent timestamp %d",
+			errInvalidTimestamp,
 			timestamp,
 			parent.Time,
 		)
@@ -232,30 +225,4 @@ func selectBigWithinBounds(lowerBound, value, upperBound *big.Int) *big.Int {
 	default:
 		return value
 	}
-}
-
-func parseFeeWindow(bytes []byte) (ap3.Window, error) {
-	if len(bytes) < FeeWindowSize {
-		return ap3.Window{}, fmt.Errorf("%w: expected at least %d bytes but got %d bytes",
-			errDynamicFeeWindowInsufficientLength,
-			FeeWindowSize,
-			len(bytes),
-		)
-	}
-
-	var window ap3.Window
-	for i := range window {
-		offset := i * wrappers.LongLen
-		window[i] = binary.BigEndian.Uint64(bytes[offset:])
-	}
-	return window, nil
-}
-
-func feeWindowBytes(w ap3.Window) []byte {
-	bytes := make([]byte, FeeWindowSize)
-	for i, v := range w {
-		offset := i * wrappers.LongLen
-		binary.BigEndian.PutUint64(bytes[offset:], v)
-	}
-	return bytes
 }