+++ RELEASES.md
@@ -1,8 +1,11 @@
 # Release Notes
 
 ## Pending Release
+
 - Bump golang version to v1.23.6
 - Bump golangci-lint to v1.63 and add linters
+- Implement ACP-176
+- Add `GasTarget` to the chain config to allow modifying the chain's `GasTarget` based on the ACP-176 rules
 
 - Added `eth_suggestPriceOptions` API to suggest gas prices (slow, normal, fast) based on the current network conditions
 - Added `"po-slow-fee-percentage"`, `"po-fast-fee-percentage"`, `"po-max-base-fee"`, and `"po-max-tip"` config flags to configure the new `eth_suggestPriceOptions` API